+++ src/components/LoginForm/LoginForm.module.css
@@ -1,0 +1,177 @@
+.container {
+    width: 280px;
+    background-color: var(--color-white);
+    box-shadow: 3px 8px 14px rgba(136, 198, 253, 0.19);
+    border-radius: 20px;
+    margin-left: auto;
+    margin-right: auto;
+    padding-right: 12px;
+    padding-left: 12px;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    position: relative;
+}
+
+@media screen and (min-width:768px) {
+    .container {
+        width: 608px;
+        padding-left: 75px;
+        padding-right: 75px;
+    }
+}
+
+
+
+.header {
+    color: var(--color-black);
+    width: 140px;
+    padding: 40px 70px;
+    font-style: normal;
+    font-weight: 500;
+    font-size: 24px;
+    margin: 0;
+    display: inline-flex;
+    align-items: center;
+    justify-content: center;
+}
+
+@media screen and (min-width:768px) {
+    .header {
+        width: 226px;
+        padding: 60px 191px 40px 191px;
+        font-size: 36px;
+
+    }
+}
+
+.formRegister {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    gap: 24px;
+}
+
+@media screen and (min-width:768px) {
+    .formRegister {
+        gap: 32px;
+
+    }
+}
+
+.inputRegister {
+    width: 256px;
+    height: 48px;
+    border: 1px solid var(--color-blue);
+    border-radius: 40px;
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    padding: 12px 16px;
+    color: var(--color-grey);
+    outline: none;
+
+}
+
+.inputRegister:focus {
+    border: 3px solid var(--color-gradient_blue);
+}
+
+@media screen and (min-width:768px) {
+    .inputRegister {
+        width: 458px;
+
+    }
+}
+
+.bttnRegister {
+    display: flex;    
+    background-color: var(--text-color-blue);
+    color: var(--color-white);
+    padding: 10px 100px;
+    flex-direction: row;
+    align-items: center;
+    justify-content: center;
+    font-style: normal;
+    font-weight: 600;
+    font-size: 20px;
+    line-height: 1.35;
+    width: 256px;
+    border-radius: 40px;
+    border-color: transparent;
+    cursor: pointer;
+    margin-top: 36px;
+}
+
+@media screen and (min-width:768px) {
+    .bttnRegister {
+        width: 458px;
+        margin-top: 28px;
+    }
+}
+
+
+.containerLog {
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    align-items: center;
+    margin-top: 24px;
+    margin-bottom: 40px;
+}
+
+@media screen and (min-width:768px) {
+    .containerLog {
+        margin-top: 16px;
+        margin-bottom: 60px;
+
+    }
+}
+
+.text {
+    font-style: normal;
+    font-weight: 400;
+    font-size: 12px;
+    line-height: 1.33;
+    color: var(--color-grey);
+    margin: 0;
+}
+
+.link {
+    font-style: normal;
+    font-weight: 400;
+    font-size: 12px;
+    line-height: 1.33;
+    color: var(--text-color-blue)
+}
+
+.passwordCont {
+    position: relative;
+}
+
+.eyeBttn {
+    border: none;
+    background-color: transparent;
+    padding: 0;
+    position: absolute;
+    top: 12px;
+    left: 216px;
+}
+
+@media screen and (min-width: 768px) {
+    .eyeBttn {
+        left: 420px;
+    }
+}
+
+
+.errorRassword {
+    margin: 0;
+    position: absolute;
+    top: 55px;
+    left: 10px;
+    color: red;
+    font-size: 12px;
+}
