+++ src/pages/LoginPage/LoginPage.jsx
@@ -1,6 +1,8 @@
+import LoginForm from "../../components/LoginForm/LoginForm";
+
 const LoginPage = () => {
     return (
-        <div></div>
+        <LoginForm></LoginForm>
     );
 }
